+++ src/components/EditableField/__tests__/EditableField.Input.test.tsx
@@ -1,8 +1,6 @@
 import * as React from 'react'
-import { cy } from '@helpscout/cyan'
 import { mount } from 'enzyme'
 import { EditableFieldInput } from '../EditableField.Input'
-import { TRUNCATED_CLASSNAMES } from '../EditableField.utils'
 
 describe('Should component update', () => {
   test('fieldValue', () => {
@@ -73,30 +71,5 @@ describe('Should component update', () => {
         actualState
       )
     ).toBeTruthy()
-  })
-})
-
-describe('email type', () => {
-  test('should truncate', () => {
-    const val = {
-      value: 'email@somethingcool.com',
-      id: 'email_0',
-    }
-
-    cy.render(<EditableFieldInput type="email" name="email" fieldValue={val} />)
-
-    expect(
-      cy.get(`.${TRUNCATED_CLASSNAMES.withSplitter}`).exists()
-    ).toBeTruthy()
-    expect(cy.get(`.${TRUNCATED_CLASSNAMES.firstChunk}`).exists()).toBeTruthy()
-    expect(
-      cy.get(`.${TRUNCATED_CLASSNAMES.splitterChunk}`).exists()
-    ).toBeTruthy()
-    expect(cy.get(`.${TRUNCATED_CLASSNAMES.secondChunk}`).exists()).toBeTruthy()
-    expect(cy.get(`.${TRUNCATED_CLASSNAMES.firstChunk}`).text()).toBe('email')
-    expect(cy.get(`.${TRUNCATED_CLASSNAMES.splitterChunk}`).text()).toBe('@')
-    expect(cy.get(`.${TRUNCATED_CLASSNAMES.secondChunk}`).text()).toBe(
-      'somethingcool.com'
-    )
   })
 })
